Add jwt token and authorize protection

+++ Cefalo.InfedgeBlog.Api/Controllers/UserController.cs
@@ -1,6 +1,8 @@
 ﻿using Cefalo.InfedgeBlog.Database.Models;
+using Cefalo.InfedgeBlog.Service.CustomExceptions;
 using Cefalo.InfedgeBlog.Service.Dtos;
 using Cefalo.InfedgeBlog.Service.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Cefalo.InfedgeBlog.Api.Controllers
@@ -10,9 +12,11 @@ namespace Cefalo.InfedgeBlog.Api.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserService _userService;
-        public UserController(IUserService userService)
+        private readonly IAuthService _authService;
+        public UserController(IUserService userService, IAuthService authService)
         {
             _userService = userService;
+            _authService = authService;
         }
         [HttpGet]
         public async Task<ActionResult<IEnumerable<User>>> GetUsersAsync()
@@ -25,21 +29,35 @@ namespace Cefalo.InfedgeBlog.Api.Controllers
             var user = await _userService.GetUserByIdAsync(Id);
             return Ok(user);
         }
-        [HttpPost]
+        [HttpPost, Authorize]
         public async Task<IActionResult> PostUserAsync([FromBody] UserPostDto userPostDto)
         {
             var userDto = await _userService.PostUserAsync(userPostDto);
+            if(userDto== null) 
+            {
+                return BadRequest("Can not create user");
+            }
             return Created("", userDto);
         }
-        [HttpPut("{Id}")]
+        [HttpPut("{Id}"), Authorize]
         public async Task<IActionResult> UpdateUserByIdAsync(int Id, [FromBody] UserUpdateDto userUpdateDto)
         {
+            var loggedInUserId = _authService.GetLoggedInUserId();
+            if(loggedInUserId != Id)
+            {
+                throw new UnauthorizedException("You are not authorized to perform this action.");
+            }
             var userDto = await _userService.UpdateUserByIdAsync(Id, userUpdateDto);
             return Ok(userDto);
         }
-        [HttpDelete("{Id}")]
+        [HttpDelete("{Id}"), Authorize]
         public async Task<IActionResult> DeleteUserByIdAsync(int Id)
         {
+            var loggedInUserId = _authService.GetLoggedInUserId();
+            if (loggedInUserId != Id)
+            {
+                throw new UnauthorizedException("You are not authorized to perform this action.");
+            }
             await _userService.DeleteUserByIdAsync(Id);
             return NoContent();
         }
